feat: 修改id类型

+++ models/server.go
@@ -4,7 +4,7 @@ import "time"
 
 // Server 表示服务器表的结构体
 type Server struct {
-	ServerID      int64     `json:"server_id" gorm:"primaryKey"`
+	ServerID      uint      `json:"server_id" gorm:"primaryKey"`
 	ServerName    string    `json:"server_name" form:"server_name" validate:"required,min=2,max=20"`
 	CreatorUserID uint      `json:"creator_user_id" validate:"required"`
 	ServerType    string    `json:"server_type" form:"server_type"`
